admin_api/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-filter==22.1
psycopg2-binary==2.9.5
gunicorn==20.1.0
python-decouple==3.7.0
redis==4.3.1
celery==5.2.7
pika==1.3.2